requirements.txt
PyMySQL==0.10.1
twython==3.4.0
python-dateutil==2.9.0.post0
pyproj==3.7.2

# transitive
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
oauthlib==3.3.1
packaging==25.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
urllib3==2.5.0
wheel==0.46.1